cpu-o3,mem-ruby,misc: port InvisiSpec, drop Futuristic mode
Ports InvisiSpec (MICRO'18) into gem5's O3 CPU + Ruby memory system.
Fixes bugs and crashes that were breaking coherence-state security and
kept the Spectre PoCs from running to completion. Removes Futuristic-
mode support (Spectre mode only, going forward). Makes the per-core
Speculative Buffer and its Load Queue effectively unbounded in size,
for simulation purposes.

+++ src/arch/arm/ArmSystem.py
@@ -40,8 +40,8 @@ from m5.params import *
 from m5.SimObject import *
 from m5.util.fdthelper import *
 
-from System import System
-from ArmSemihosting import ArmSemihosting
+from m5.objects.System import System
+from m5.objects.ArmSemihosting import ArmSemihosting
 
 class ArmMachineType(Enum):
     map = {
@@ -95,7 +95,7 @@ class ArmSystem(System):
         # root instead of appended.
 
         def generateMemNode(mem_range):
-            node = FdtNode("memory@%x" % long(mem_range.start))
+            node = FdtNode("memory@%x" % int(mem_range.start))
             node.append(FdtPropertyStrings("device_type", ["memory"]))
             node.append(FdtPropertyWords("reg",
                 state.addrCells(mem_range.start) +
